refactor: little changes after review
