+++ .eslintrc.json
@@ -27,6 +27,7 @@
     "@typescript-eslint/comma-dangle": "warn",
     "@typescript-eslint/strict-boolean-expressions": "off",
     "@typescript-eslint/object-curly-spacing": "warn",
-    "@typescript-eslint/explicit-function-return-type": "off"
+    "@typescript-eslint/explicit-function-return-type": "off",
+    "@typescript-eslint/no-invalid-void-type": "off"
   }
 }
