+++ src/api.ts
@@ -1,6 +1,6 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
 
-import { type FilmShort, type Response } from './types/Film'
+import { type Response } from './types/types'
 
 const requestHeaders = {
   'X-API-KEY': process.env.REACT_APP_KINOPOISK_API_KEY,
@@ -27,7 +27,6 @@ export const filmAPI = createApi({
     getFilmsByKeyword: build.query({
       query: (keyword: string) => createRequest(`v1.2/movie/search?page=1&limit=5&query=${keyword}`)
     }),
-    // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
     getTop24: build.query<Response, void>({
       query: () => createRequest('v1.3/movie?page=1&limit=24&selectFields=id&selectFields=name&selectFields=description&selectFields=year&selectFields=poster')
     })
